Move Card components to new editor
