+++ app/src/docs/Cards/index.mdx
@@ -1,0 +1,27 @@
+import { NPM } from "../lib/NPM";
+import { CardMono } from "./CardMono";
+import { CardInGrid } from "./CardInGrid";
+import { CardWithGridSystem } from "./CardWithGridSystem";
+import { CardWithGridSystemTwoRows } from "./CardWithGridSystemTwoRows";
+
+# Cards
+
+## Simple card components.
+
+<NPM npm={props.npm} />
+
+### Simple Card
+
+<CardMono />
+
+### Card in Grid system - 1 row - 3 columns, with different animations
+
+<CardInGrid />
+
+### Card with included Grid system - 1 row - 3 columns
+
+<CardWithGridSystem />
+
+### Card with included Grid system - 2 row - 3 columns
+
+<CardWithGridSystemTwoRows />
