+++ app/src/docs/Cards/index.tsx
@@ -1,0 +1,10 @@
+/* @license Copyright 2023 @paritytech/polkadot-cloud authors & contributors
+SPDX-License-Identifier: GPL-3.0-only */
+
+import Doc from "./index.mdx";
+
+export const Card = () => (
+  <div className="doc">
+    <Doc npm="@polkadot-cloud/react" />
+  </div>
+);
