--- app/src/docs/cards.mdx
@@ -1,7 +1,0 @@
-import { NPM } from "./lib/NPM";
-
-# Cards
-
-## Simple card components.
-
-<NPM npm={props.npm} />
